app/requirements.txt
python-bittrex
pymongo
requests
python-binance